Song page is working again
- Update file naming convention (Part 5)
